moved grpc methods to .grpc files accordingly. pr #96
